+++ rotkehlchen/chain/ethereum/defi.py
@@ -31,6 +31,7 @@ CURVEFI_RENSWAP = EthereumConstants().contract('CURVEFI_RENSWAP')
 CURVEFI_SRENSWAP = EthereumConstants().contract('CURVEFI_SRENSWAP')
 CURVEFI_SUSDV2SWAP = EthereumConstants().contract('CURVEFI_SUSDV2SWAP')
 CURVEFI_3POOLSWAP = EthereumConstants().contract('CURVEFI_3POOLSWAP')
+CURVEFI_GUSDC3CRVSWAP = EthereumConstants().contract('CURVEFI_GUSDC3CRVSWAP')
 YEARN_CONTROLLER = EthereumConstants().contract('YEARN_CONTROLLER')
 
 HARVEST_VAULTS = ('fUSDC', 'fUSDT', 'fDAI', 'fWETH', 'fTUSD', 'fWBTC', 'frenBTC', 'fcrvRenWBTC')
@@ -209,6 +210,9 @@ def handle_defi_price_query(
     elif token_symbol == '3Crv':
         token = EthereumToken(token_symbol)
         usd_value = _handle_curvepool_price(ethereum, CURVEFI_3POOLSWAP, token.decimals, ONE)
+    elif token_symbol == 'gusd3CRV':
+        token = EthereumToken(token_symbol)
+        usd_value = _handle_curvepool_price(ethereum, CURVEFI_GUSDC3CRVSWAP, token.decimals, ONE)
     elif token_symbol == 'yDAI+yUSDC+yUSDT+yBUSD':
         token = EthereumToken(token_symbol)
         usd_value = _handle_curvepool_price(ethereum, CURVEFI_BUSDSWAP, token.decimals, ONE)